Switch to expectation-style tests
In this style of testing, the result is printed to stdout rather than
compared against an expected result programatically.

The primary advantage to this style of testing in the context is that
it is easier to identify what went wrong when a test fails: the diff
highlights the mismatch.

The tests have been re-ordered so that they correspond to the order of
declarations in the .mli file.